refactor logging configuration
