+++ app/core/logger.py
@@ -2,31 +2,19 @@ import logging
 import sys
 
 from .config import settings
-
-LOG_LEVEL = getattr(logging, settings.LOG_LEVEL.upper(), logging.INFO)
-
-
-class CustomFormatter(logging.Formatter):
-    def format(self, record):
-        levelname = record.levelname
-        if levelname == 'INFO':
-            levelname = 'INFO:    '
-        elif levelname == 'DEBUG':
-            levelname = 'DEBUG:   '
-        elif levelname == 'WARNING':
-            levelname = 'WARNING: '
-        elif levelname == 'ERROR':
-            levelname = 'ERROR:   '
-        elif levelname == 'CRITICAL':
-            levelname = 'CRITICAL:'
-        record.levelname = levelname
-        return super().format(record)
+from .logging_formatter import CustomFormatter
 
 
-formatter = CustomFormatter('%(levelname)s [%(name)s] %(message)s')
+def configure_logging(level: str | None = None, stream=sys.stdout) -> None:
+    """Configure application logging."""
+    level_name = level or settings.LOG_LEVEL
+    log_level = getattr(logging, level_name.upper(), logging.INFO)
 
-logging.basicConfig(level=LOG_LEVEL, handlers=[logging.StreamHandler(sys.stdout)])
-
-
-for handler in logging.root.handlers:
+    handler = logging.StreamHandler(stream)
+    formatter = CustomFormatter('%(levelname)s [%(name)s] %(message)s')
     handler.setFormatter(formatter)
+
+    logging.basicConfig(level=log_level, handlers=[handler])
+
+    for h in logging.root.handlers:
+        h.setFormatter(formatter)
